requirements.txt
python-dotenv
click
colorama
pyyaml
requests
